// lib/api/generated/models/contactMessage.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CKS Slavia API
 * API panelu klubowego CKS Slavia Ruda Śląska
 * OpenAPI spec version: 1.1.3.0+1
 */

export interface ContactMessage {
  body: string;
  created_at: string;
  email: string;
  id: string;
  name: string;
  /** Numer w formacie międzynarodowym, np. +48 500123456. */
  phone: string;
  read: boolean;
  /** @nullable */
  read_at?: string | null;
  /** @nullable */
  read_by?: string | null;
  subject: string;
}